imported_pip_requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
deprecation==2.1.0
Django==3.2.25
django-allauth==0.62.1
django-ckeditor==6.3.2
django-clearcache==1.2.1
django-js-asset==2.0.0
django-rest-passwordreset==1.4.1
django-rest-resetpassword==0.1.2
djangorestframework==3.15.1
oauthlib==3.2.2
packaging==24.0
Pillow==9.5.0
pycparser==2.21
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1